+++ src/config/defaultConfig.js
@@ -10,9 +10,9 @@ const defaultConifg = {
   fetchOptions: {
     headers: {
       Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-  },
-};
+      'Content-Type': 'application/json'
+    }
+  }
+}
 
-export default defaultConifg;
+export default defaultConifg